requirements.txt
google-cloud-firestore
pytz # For robust timezone handling in Python